Mongo EnumEnvironment + getdatabase(enumEnv..)
