+++ src/main/java/ulaval/glo2003/floppa/app/mongoRepository/MongoEnumEnvironment.java
@@ -1,0 +1,11 @@
+package ulaval.glo2003.floppa.app.mongoRepository;
+
+public class MongoEnumEnvironment {
+
+    enum environment {
+        PRODUCTION,
+        STAGING,
+        TEST
+
+    }
+}
